backend-services/python/src/flask/requirements.txt
flask==3.0.1
Authlib==1.3.0
python-dotenv==1.0.1
